requirements.txt
fastapi>=0.109.0
uvicorn>=0.35.0
openai>=1.6.0
pinecone-client>=5.0.1
python-dotenv>=1.0.0